[SQUASHED] Settings: Heads up meets SlimKat (4/4)
For detailed explanation see here: https://gerrit.slimroms.net/#/c/9937/

This patch adds the 4 possible user configurable settings.
1. Enable always expanded heads up notifications.
2. Set heads up snooze time
3. Set timout timer
4. Show updates of pending notifications
5. Add in app info screen a checkbox to enable heads up for the app.
   As another possibility aside the longpress on the notification on
   notification drawer.

Conflicts:
	res/values/cr_arrays.xml
	res/values/cr_strings.xml
	res/xml/settings_headers.xml
	src/com/android/settings/applications/InstalledAppDetails.java

Settings: Show heads up at the bottom of the screen (2/2)
Due of user requests and indeed especially on hw key devices for a better
workflow we add an option to show heads up notifications on the bottom instead
of the top.

Animation we leave as is due that if it slide from the bottom into the screen it happens
to often that the user accidentaly interacts with it. If it comes from the top of the
container we add, the user has enough time to realize it. Sometimes a ms can make a big difference.

Heads up: enable heads up notification checkbox for all

- Disbale check of system packages to allow checkbox for all. This will, for example, bring screenshot to heads up if SystemUI is enabled

Change-Id: I273caffa992e79a9a9d6914bf1c89f9388044da4

Heads up: option to disable in lockscreen (2/2)

Change-Id: Icbc6fba1727c009f7ea42a66ea3167f19d527e10

Conflicts:
	res/values/cr_strings.xml